Added query cleanup, renamed package
